+++ src/matmodules.cpp
@@ -342,64 +342,6 @@ bool MOutput::Initialize() {
 
 
 /*********************
-matrix Product module.
-**********************/
-MProduct::~MProduct() {}
-BusSize MProduct::Get_OutputBusSize() const { return _size; }
-u8 MProduct::Get_State() const { return _state; }
-MProduct::MProduct(Simulator *sim, std::string name): MatModule(sim, name) {
-    _state = 0;
-    _portcnt = 0;
-    MATMODULE_INIT();
-}
-PUnitModule MProduct::Get_OutputPort(BusSize size) const {
-    if (_misoy==nullptr) TRACELOG(LOG_FATAL, "internal error: MProduct.");
-    if (!(size<_size)) return nullptr;
-    return _misoy[size.r*_size.c+size.c];
-}
-bool MProduct::Initialize() {
-    if (_state == BUS_INITIALIZED) return true;
-    if (_nextL==nullptr) TRACELOG(LOG_FATAL, "MProduct: \"%s\" doesn't have 2 child modules!", _name.c_str());
-    if (!(_nextL->Get_State() & BUS_SIZED)) return false;
-    if (!(_nextR->Get_State() & BUS_SIZED)) return false;
-    _size = _nextL->Get_OutputBusSize();
-    BusSize childSize = _nextR->Get_OutputBusSize();
-    if (_size.c != childSize.r)
-        TRACELOG(LOG_FATAL, "MProduct: Bus size mismatch between child modules of \"%s\"!\n    "
-        "left:%d,%d; right:%d,%d", _name.c_str(), _size.r, _size.c, childSize.r, childSize.c);
-    _size.c = childSize.c;
-    _misoy = new PUFcnMISO[_size.r*_size.c];
-    for (uint i = 0; i < _size.r; i++)
-        for (uint j = 0; j < _size.c; j++)
-            _misoy[i*_size.c+j] = new UFcnMISO(_sim, _name+"_misoy_"+std::to_string(i)+"_"+std::to_string(j));
-    _state = BUS_GENERATED;
-    if (!(_nextL->Get_State() & BUS_GENERATED)) return false;
-    if (!(_nextR->Get_State() & BUS_GENERATED)) return false;
-    for (uint i = 0; i < _size.r; i++) {
-        for (uint j = 0; j < _size.c; j++) {
-            _misoy[i*_size.c+j]->Set_Function([childSize](double *u){
-                double ans = 0;
-                for (int i = 0; i < childSize.r; i++)
-                    ans += u[i+i]*u[i+i+1];
-                return ans;
-            });
-            for (uint k=0; k<childSize.r; ++k) {
-                _sim->connectU(_nextL->Get_OutputPort(BusSize(i, k)), _misoy[i*_size.c+j]);
-                _sim->connectU(_nextR->Get_OutputPort(BusSize(k, j)), _misoy[i*_size.c+j]);
-            }
-        }
-    }
-    _state = BUS_INITIALIZED; return true;
-}
-void MProduct::connect(const PMatModule m) {
-    if (_portcnt==0) _nextR = m;
-    else if (_portcnt==1) _nextL = m;
-    else TRACELOG(LOG_WARNING, "MProduct: \"%s\" is repeatedly connected.", _name.c_str());
-    _portcnt++;
-}
-
-
-/*********************
 matrix Sum module.
 **********************/
 MSum::~MSum() {}
@@ -456,33 +398,6 @@ void MSum::Set_InputGain(double inputgain, int port) {
             TRACELOG(LOG_WARNING, "MSum: \"%s\" doesn't have enough child module.", _name.c_str());
         _ingain[port] = inputgain;
     }
-}
-
-
-/*********************
-matrix MTranspose module.
-**********************/
-MTranspose::~MTranspose() {}
-BusSize MTranspose::Get_OutputBusSize() const { return _size; }
-u8 MTranspose::Get_State() const { return _state; }
-void MTranspose::connect(const PMatModule m) { _next = m; }
-MTranspose::MTranspose(Simulator *sim, std::string name): MatModule(sim, name) {
-    _state = 0;
-    MATMODULE_INIT();
-}
-PUnitModule MTranspose::Get_OutputPort(BusSize size) const {
-    if (!(size<_size)) return nullptr;
-    return _next->Get_OutputPort(BusSize(size.c, size.r));
-}
-bool MTranspose::Initialize() {
-    if (_state == BUS_INITIALIZED) return true;
-    if (_next==nullptr) TRACELOG(LOG_FATAL, "MTranspose: \"%s\" doesn't have a child module!", _name.c_str());
-    if (!(_next->Get_State() & BUS_SIZED)) return false;
-    _size = _next->Get_OutputBusSize();
-    _size = BusSize(_size.c, _size.r);
-    _state = BUS_SIZED;
-    if (!(_next->Get_State() & BUS_INITIALIZED)) return false;
-    _state = BUS_INITIALIZED; return true;
 }